Move models into their own folder, create a collection model, add type to Field type and clean up some bugs
